+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Infos/Badges/StyleInfoBadge.cs
@@ -5,7 +5,7 @@ using osu.Framework.Allocation;
 using osu.Game.Graphics;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Badges
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Infos.Badges
 {
     public class StyleInfoBadge : Badge
     {
