+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Lyrics/DrawableEditorLyric.cs
@@ -7,7 +7,7 @@ using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Objects.Drawables;
 using osu.Game.Rulesets.Karaoke.Skinning.Components;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Lyrics
 {
     public class DrawableEditLyric : DrawableLyric
     {
