+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Lyrics/TimeTags/DrawableTimeTag.cs
@@ -5,23 +5,27 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
+using osu.Framework.Input.Events;
 using osu.Game.Graphics;
 using osu.Game.Rulesets.Karaoke.Graphics.Shapes;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.TimeTags
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Lyrics.TimeTags
 {
-    public class DrawableTimeTagCursor : CompositeDrawable
+    public class DrawableTimeTag : CompositeDrawable
     {
         /// <summary>
         /// Height of major bar line triangles.
         /// </summary>
-        private const float triangle_width = 4;
+        private const float triangle_width = 3;
 
         private readonly TimeTag timeTag;
 
-        public DrawableTimeTagCursor(TimeTag timeTag)
+        [Resolved(canBeNull: true)]
+        private TimeTagManager timeTagManager { get; set; }
+
+        public DrawableTimeTag(TimeTag timeTag)
         {
             this.timeTag = timeTag;
             AutoSizeAxes = Axes.Both;
@@ -36,10 +40,15 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.TimeTags
             };
         }
 
+        protected override bool OnClick(ClickEvent e)
+        {
+            return timeTagManager?.MoveCursorToTargetPosition(timeTag) ?? false;
+        }
+
         [BackgroundDependencyLoader]
         private void load(OsuColour colours)
         {
-            InternalChild.Colour = timeTag.Time.HasValue ? colours.YellowDarker : colours.Gray3;
+            InternalChild.Colour = timeTag.Time.HasValue ? colours.Yellow : colours.Gray7;
         }
     }
 }
